+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- Upgraded to [@cucumber/language-server 0.10.3](https://github.com/cucumber/language-server/blob/main/CHANGELOG.md#0103---2022-05-12)
+
 ## [1.0.7] - 2022-05-12
 ### Added
 - Publish to Open VSX Registry as well as VS Code Marketplace